feat(llm): reorganize logic of the analytics opt in prompt
