+++ apps/ledger-live-mobile/src/components/RootNavigator/AnalyticsOptInPromptNavigator.tsx
@@ -3,13 +3,16 @@ import { createStackNavigator } from "@react-navigation/stack";
 import { useTheme } from "styled-components/native";
 import { useFeature } from "@ledgerhq/live-common/featureFlags/index";
 import { ABTestingVariants } from "@ledgerhq/types-live";
-import { ScreenName } from "~/const";
+import { NavigatorName, ScreenName } from "~/const";
 import { getStackNavigatorConfig } from "~/navigation/navigatorConfig";
 import AnalyticsOptInPromptMainA from "~/screens/AnalyticsOptInPrompt/variantA/Main";
 import AnalyticsOptInPromptDetailsA from "~/screens/AnalyticsOptInPrompt/variantA/Details";
 import AnalyticsOptInPromptMainB from "~/screens/AnalyticsOptInPrompt/variantB/Main";
 import AnalyticsOptInPromptDetailsB from "~/screens/AnalyticsOptInPrompt/variantB/Details";
 import { AnalyticsOptInPromptNavigatorParamList } from "./types/AnalyticsOptInPromptNavigator";
+import { useRoute } from "@react-navigation/core";
+import { RootComposite, StackNavigatorProps } from "./types/helpers";
+import { BaseNavigatorStackParamList } from "./types/BaseNavigator";
 
 const screensByVariant = {
   [ABTestingVariants.variantA]: {
@@ -21,6 +24,9 @@ const screensByVariant = {
     details: AnalyticsOptInPromptDetailsB,
   },
 };
+type NavigationProps = RootComposite<
+  StackNavigatorProps<BaseNavigatorStackParamList, NavigatorName.AnalyticsOptInPrompt>
+>;
 
 const Stack = createStackNavigator<AnalyticsOptInPromptNavigatorParamList>();
 
@@ -28,6 +34,13 @@ export default function AnalyticsOptInPromptNavigator() {
   const { colors } = useTheme();
   const stackNavigationConfig = useMemo(() => getStackNavigatorConfig(colors, false), [colors]);
   const llmAnalyticsOptInPromptFeature = useFeature("llmAnalyticsOptInPrompt");
+  const route = useRoute<NavigationProps["route"]>();
+  const preventBackNavigation = route.params.params?.entryPoint === "Portfolio";
+
+  const navigationOptions = {
+    title: "",
+    ...(preventBackNavigation ? { headerLeft: () => null } : {}),
+  };
 
   const activeVariant =
     llmAnalyticsOptInPromptFeature?.params?.variant === ABTestingVariants.variantB
@@ -39,7 +52,7 @@ export default function AnalyticsOptInPromptNavigator() {
       <Stack.Screen
         name={ScreenName.AnalyticsOptInPromptMain}
         component={screensByVariant[activeVariant].main}
-        options={{ title: "" }}
+        options={navigationOptions}
       />
       <Stack.Screen
         name={ScreenName.AnalyticsOptInPromptDetails}
