+++ apps/ledger-live-mobile/src/screens/AnalyticsOptInPrompt/variantA/Details.tsx
@@ -6,7 +6,7 @@ import { useTranslation } from "react-i18next";
 import Button from "~/components/Button";
 import Switch from "~/components/Switch";
 import { View, Container, Titles, Content, Bottom, ScrollableContainer } from "../Common";
-import useAnalyticsOptInPrompt from "~/hooks/useAnalyticsOptInPromptVariantA";
+import useAnalyticsOptInPrompt from "~/hooks/analyticsOptInPrompt/useAnalyticsOptInPromptVariantA";
 import { StackNavigatorProps } from "~/components/RootNavigator/types/helpers";
 import { AnalyticsOptInPromptNavigatorParamList } from "~/components/RootNavigator/types/AnalyticsOptInPromptNavigator";
 import { ScreenName } from "~/const";
